[refactor/postsNetworkBoundResource] refactor: change post ID type from Int to Long in PostDao for better compatibility

+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/data/database/PostDao.kt
@@ -24,5 +24,5 @@ internal interface PostDao {
     suspend fun deleteAllPosts()
 
     @Query("DELETE FROM posts WHERE id = :postId")
-    suspend fun deletePostById(postId: Int)
+    suspend fun deletePostById(postId: Long)
 }
